ios [nfc]: Move IosNativeHostApiImpl to its own Swift file
This commit is a result of opening context menu on the class
in Xcode's editor and then selecting "Refactor > Extract to file".

This is needed for (soon to be added) NotificationService target to
initialize the Pigeon API.
