+++ ios/Runner/AppDelegate.swift
@@ -87,13 +87,3 @@ class NotificationTapEventListener: NotificationTapEventsStreamHandler {
     eventSink?.success(IosNotificationTapEvent(payload: payload))
   }
 }
-
-private class IosNativeHostApiImpl: IosNativeHostApi {
-  func setExcludedFromBackup(filePath: String) throws {
-    var resourceValues = URLResourceValues()
-    resourceValues.isExcludedFromBackup = true
-
-    var url = URL(fileURLWithPath: filePath, isDirectory: false)
-    try url.setResourceValues(resourceValues)
-  }
-}
