+++ ios/Runner/IosNativeHostApi.swift
@@ -1,0 +1,12 @@
+import Flutter
+import UIKit
+
+public class IosNativeHostApiImpl: IosNativeHostApi {
+  func setExcludedFromBackup(filePath: String) throws {
+    var resourceValues = URLResourceValues()
+    resourceValues.isExcludedFromBackup = true
+
+    var url = URL(fileURLWithPath: filePath, isDirectory: false)
+    try url.setResourceValues(resourceValues)
+  }
+}
